[ui] cleanup, parseLogMessage
