+++ frontend/src/DeviceInfo.js
@@ -1,0 +1,5 @@
+export default {
+  initDevice: () => {
+    console.log('init web device')
+  }
+}
